ruff check --select RUF --fix

+++ hail/python/hail/ggplot/facets.py
@@ -103,7 +103,7 @@ class FacetWrap(Faceter):
         elif self.nrow is not None:
             return (self.nrow, n_partitions(num_facet_values, self.nrow))
         else:
-            ncol = int(math.ceil(math.sqrt(num_facet_values)))
+            ncol = math.ceil(math.sqrt(num_facet_values))
             return (n_partitions(num_facet_values, ncol), ncol)
 
     def get_shared_axis_kwargs(self) -> Dict[str, str]:
